+++ CHANGELOG.md
@@ -29,6 +29,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Deprecations
 
+* The `all:` revset modifier and `ui.always-allow-large-revsets` setting is
+  planned to be removed in a future release.
+  [#6016](https://github.com/jj-vcs/jj/issues/6016)
+
 ### New features
 
 * `jj workspace list` now accepts `-T`/`--template` option to customize its output via templates.